refactor: move api models to own folder
