+++ src/depot_server/api/models/bays.py
@@ -1,7 +1,8 @@
-from authlib.oidc.core import UserInfo
-from fastapi import APIRouter, Depends, Body, HTTPException, Response
 from typing import List
 from uuid import UUID, uuid4
+
+from authlib.oidc.core import UserInfo
+from fastapi import APIRouter, Depends, Body, HTTPException, Response
 
 from depot_server.db import DbBay, collections
 from depot_server.helper.auth import Authentication
