+++ src/depot_server/api/models/items.py
@@ -1,19 +1,20 @@
-from authlib.oidc.core import UserInfo
-from fastapi import APIRouter, Depends, Body, Query, HTTPException, BackgroundTasks, Response
 from typing import Annotated, List, Optional, Dict
 from uuid import UUID, uuid4
+
+from authlib.oidc.core import UserInfo
+from fastapi import APIRouter, Depends, Body, Query, HTTPException, BackgroundTasks, Response
 from pymongo import DESCENDING
 
 from depot_server.db import collections, DbItem, DbItemState, DbStrChange, \
     DbItemStateChanges, DbItemConditionChange, DbDateChange, DbIdChange, DbTagsChange, DbTotalReportStateChange, \
     DbItemReport
+from depot_server.db.model import DbReportElement
 from depot_server.helper.auth import Authentication
 from depot_server.helper.util import utc_now
+from depot_server.mail.reservation_item_removed import send_reservation_item_removed
 from depot_server.model import Item, ItemInWrite, ReportItemInWrite, ItemCondition, ReservationState
+from depot_server.model.item_state import ItemReport
 from depot_server.model.reservation import Reservation
-from ..db.model import DbReportElement
-from ..mail.reservation_item_removed import send_reservation_item_removed
-from ..model.item_state import ItemReport
 
 router = APIRouter()
