+++ src/depot_server/api/models/pictures.py
@@ -1,14 +1,12 @@
+import hashlib
 from io import BytesIO
+from typing import List, cast, Tuple
 
 import gridfs
-import hashlib
+from PIL import Image
 from authlib.oidc.core import UserInfo
 from fastapi import APIRouter, Depends, HTTPException, UploadFile, File, Header, Response
 from starlette.responses import StreamingResponse
-from typing import List, cast, Tuple
-
-from PIL import Image
-
 
 from depot_server.db import collections
 from depot_server.helper.auth import Authentication
