+++ src/depot_server/api/models/reservations.py
@@ -1,10 +1,11 @@
 import random
-from authlib.oidc.core import UserInfo
 from datetime import date
-from fastapi import APIRouter, Depends, Body, Query, HTTPException, BackgroundTasks, Response
-from pymongo import DESCENDING, ASCENDING
 from typing import List, Optional, Set, Dict
 from uuid import UUID, uuid4
+
+from authlib.oidc.core import UserInfo
+from fastapi import APIRouter, Depends, Body, Query, HTTPException, BackgroundTasks, Response
+from pymongo import DESCENDING, ASCENDING
 
 from depot_server.config import config
 from depot_server.db import collections, DbReservation, DbItem, DbItemReservation
